[Gold III] Title: 당근 훔쳐 먹기, Time: 636 ms, Memory: 67656 KB -BaekjoonHub

+++ 백준/Gold/18234. 당근 훔쳐 먹기/당근 훔쳐 먹기.java
@@ -43,7 +43,7 @@ public class Main {
 		Collections.sort(list);
 		for(int i = T-N; i < T; i++) {
 			Carrot carrot = list.get(list.size() - 1);
-			result += (long)( carrot.w + (long)(carrot.p * (long)i));
+			result += ( carrot.w + (carrot.p * (long)i));
 			list.remove(list.size() - 1);
 		}
 		System.out.println(result);
